fix: gridworld mdp impl

+++ src/rlplg/learning/tabular/dynamicprog.py
@@ -22,7 +22,7 @@ def iterative_policy_evaluation(
     Vectorized implementation.
     """
     pi_action = np.zeros(
-        (mdp.env_desc().num_states, mdp.env_desc().num_actions), dtype=np.float32
+        (mdp.env_desc().num_states, mdp.env_desc().num_actions), dtype=np.float64
     )
     transition = np.zeros(
         (
@@ -30,7 +30,7 @@ def iterative_policy_evaluation(
             mdp.env_desc().num_actions,
             mdp.env_desc().num_states,
         ),
-        dtype=np.float32,
+        dtype=np.float64,
     )
     reward = np.zeros(
         (
@@ -38,7 +38,7 @@ def iterative_policy_evaluation(
             mdp.env_desc().num_actions,
             mdp.env_desc().num_states,
         ),
-        dtype=np.float32,
+        dtype=np.float64,
     )
 
     for state in range(mdp.env_desc().num_states):
